[refactor] PointController 분리 (#440)
