+++ backend/src/main/java/dev/tripdraw/trip/presentation/TripController.java
@@ -7,9 +7,6 @@ import dev.tripdraw.common.auth.Auth;
 import dev.tripdraw.common.auth.LoginUser;
 import dev.tripdraw.common.swagger.SwaggerAuthorizationRequired;
 import dev.tripdraw.trip.application.TripService;
-import dev.tripdraw.trip.dto.PointCreateRequest;
-import dev.tripdraw.trip.dto.PointCreateResponse;
-import dev.tripdraw.trip.dto.PointResponse;
 import dev.tripdraw.trip.dto.TripCreateResponse;
 import dev.tripdraw.trip.dto.TripResponse;
 import dev.tripdraw.trip.dto.TripSearchRequest;
@@ -27,7 +24,6 @@ import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @Tag(name = "Trip", description = "여행 관련 API 명세")
@@ -49,35 +45,6 @@ public class TripController {
         return ResponseEntity.status(CREATED).body(response);
     }
 
-    @Operation(summary = "위치 정보 저장 API", description = "현재 진행 중인 여행의 경로에 위치 정보를 저장합니다.")
-    @ApiResponse(
-            responseCode = "201",
-            description = "위치 정보 저장 성공."
-    )
-    @PostMapping("/points")
-    public ResponseEntity<PointCreateResponse> addPoint(
-            @Auth LoginUser loginUser,
-            @RequestBody PointCreateRequest pointCreateRequest
-    ) {
-        PointCreateResponse response = tripService.addPoint(loginUser, pointCreateRequest);
-        return ResponseEntity.status(CREATED).body(response);
-    }
-
-    @Operation(summary = "위치 정보 조회 API", description = "위치 정보를 조회합니다.")
-    @ApiResponse(
-            responseCode = "200",
-            description = "위치 정보 조회 성공."
-    )
-    @GetMapping("/points/{pointId}")
-    public ResponseEntity<PointResponse> readPointById(
-            @Auth LoginUser loginUser,
-            @PathVariable Long pointId,
-            @RequestParam Long tripId
-    ) {
-        PointResponse response = tripService.readPointByTripAndPointId(loginUser, tripId, pointId);
-        return ResponseEntity.ok(response);
-    }
-
     @Operation(summary = "나의 여행 조회 API", description = "회원 한 명의 단일 여행 정보를 조회합니다.")
     @ApiResponse(
             responseCode = "200",
@@ -87,21 +54,6 @@ public class TripController {
     public ResponseEntity<TripResponse> readById(@Auth LoginUser loginUser, @PathVariable Long tripId) {
         TripResponse response = tripService.readTripById(loginUser, tripId);
         return ResponseEntity.ok(response);
-    }
-
-    @Operation(summary = "위치정보 삭제 API", description = "특정 위치정보를 삭제합니다.")
-    @ApiResponse(
-            responseCode = "204",
-            description = "위치정보 삭제 성공"
-    )
-    @DeleteMapping("/points/{pointId}")
-    public ResponseEntity<Void> deletePoint(
-            @Auth LoginUser loginUser,
-            @PathVariable Long pointId,
-            @RequestParam Long tripId
-    ) {
-        tripService.deletePoint(loginUser, pointId, tripId);
-        return ResponseEntity.noContent().build();
     }
 
     @Operation(summary = "나의 여행 전체 조회 API", description = "회원 한 명의 모든 여행 정보를 조회합니다.")
@@ -121,10 +73,7 @@ public class TripController {
             description = "모든 회원 여행 전체 조회 성공."
     )
     @GetMapping("/trips")
-    public ResponseEntity<TripsSearchResponse> readAll(
-            @Auth LoginUser loginUser,
-            TripSearchRequest tripSearchRequest
-    ) {
+    public ResponseEntity<TripsSearchResponse> readAll(TripSearchRequest tripSearchRequest) {
         TripsSearchResponse response = tripService.readAll(tripSearchRequest);
         return ResponseEntity.ok(response);
     }
